+++ InAppSettingsKit/Views/IASKPSTitleValueSpecifierViewCell.m
@@ -17,6 +17,7 @@
 #import "IASKPSTitleValueSpecifierViewCell.h"
 #import "IASKSettingsReader.h"
 
+#import <tgmath.h>
 
 @implementation IASKPSTitleValueSpecifierViewCell
 
@@ -26,7 +27,7 @@
 	CGSize viewSize =  [self.textLabel superview].frame.size;
 
 	// if there's an image, make room for it
-	CGFloat imageOffset = floorf(self.imageView.image ? self.imageView.bounds.size.width + self.imageView.frame.origin.x : 0);
+	CGFloat imageOffset = floor(self.imageView.image ? self.imageView.bounds.size.width + self.imageView.frame.origin.x : 0);
   
 	// set the left title label frame
 	CGFloat labelWidth = [self.textLabel sizeThatFits:CGSizeZero].width;
